Add 'Windows.UI.Xaml' exception types
