+++ src/WinRT.Runtime/MatchingRefApiCompatBaseline.txt
@@ -273,4 +273,8 @@ TypesMustExist : Type 'WinRT.GeneratedBindableCustomPropertyAttribute' does not 
 TypesMustExist : Type 'Microsoft.UI.Xaml.Data.IBindableCustomPropertyImplementation' does not exist in the reference but it does exist in the implementation.
 MembersMustExist : Member 'public void ABI.System.Uri.DisposeAbiArray(System.Object)' does not exist in the reference but it does exist in the implementation.
 TypesMustExist : Type 'WinRT.EnumTypeDetails<T>' does not exist in the reference but it does exist in the implementation.
-Total Issues: 274
+TypesMustExist : Type 'Windows.UI.Xaml.LayoutCycleException' does not exist in the reference but it does exist in the implementation.
+TypesMustExist : Type 'Windows.UI.Xaml.Automation.ElementNotAvailableException' does not exist in the reference but it does exist in the implementation.
+TypesMustExist : Type 'Windows.UI.Xaml.Automation.ElementNotEnabledException' does not exist in the reference but it does exist in the implementation.
+TypesMustExist : Type 'Windows.UI.Xaml.Markup.XamlParseException' does not exist in the reference but it does exist in the implementation.
+Total Issues: 278
